Refactor wagtailpages/factory.py into a module
