+++ network-api/networkapi/wagtailpages/factory/mini_site_namespace.py
@@ -1,0 +1,7 @@
+from networkapi.wagtailpages.models import MiniSiteNameSpace
+from wagtail_factories import PageFactory
+
+
+class MiniSiteNamespaceFactory(PageFactory):
+    class Meta:
+        model = MiniSiteNameSpace
