+++ network-api/networkapi/wagtailpages/factory/primary_page.py
@@ -1,0 +1,7 @@
+from networkapi.wagtailpages.models import PrimaryPage
+from .abstract import CMSPageFactory
+
+
+class PrimaryPageFactory(CMSPageFactory):
+    class Meta:
+        model = PrimaryPage
