+++ network-api/networkapi/wagtailpages/factory/signup.py
@@ -1,0 +1,7 @@
+from networkapi.wagtailpages.models import Signup
+from .abstract import CTAFactory
+
+
+class SignupFactory(CTAFactory):
+    class Meta:
+        model = Signup
